requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-widget-tweaks==1.4.3
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
django-import-export==1.2.0